Move ScriptCompiler code
