+++ unbound_script.h
@@ -4,6 +4,7 @@
 
 #ifdef __cplusplus
 
+#include <memory>
 #include "deps/include/v8-script.h"
 
 namespace v8 {
@@ -14,8 +15,14 @@ struct m_unboundScript {
   v8::Persistent<v8::UnboundScript> ptr;
 };
 
+typedef v8::ScriptCompiler::CachedData* ScriptCompilerCachedDataPtr;
+
 extern "C" {
 #else
+
+typedef struct v8ScriptCompilerCachedData v8ScriptCompilerCachedData;
+typedef const v8ScriptCompilerCachedData* ScriptCompilerCachedDataPtr;
+
 #endif
 
 typedef struct m_unboundScript m_unboundScript;
@@ -27,8 +34,14 @@ typedef struct {
   RtnError error;
 } RtnUnboundScript;
 
+typedef struct {
+  ScriptCompilerCachedDataPtr ptr;
+  const uint8_t* data;
+  int length;
+  int rejected;
+} ScriptCompilerCachedData;
+
 #ifdef __cplusplus
 }  // extern "C"
 #endif
-
 #endif
